Renaming actions directory to lowercase.
