+++ resources/assets/components/actions-temp/ThirdPartyAction/ThirdPartyAction.js
